Revert "typescript refactor"
