+++ src/main.jsx
@@ -1,0 +1,22 @@
+import React from 'react'
+import {
+  createBrowserRouter,
+  RouterProvider,
+} from "react-router-dom";
+import ErrorPage from "./routes/error";
+import ReactDOM from 'react-dom/client'
+import App from './Root.jsx'
+
+const router = createBrowserRouter([
+  {
+    path: "/",
+    element: <App />,
+    errorElement: <ErrorPage />,
+  },
+]);
+
+ReactDOM.createRoot(document.getElementById('root')).render(
+  <React.StrictMode>
+      <RouterProvider router={router} /> 
+  </React.StrictMode>,
+)
